requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
langchain-openai==0.2.5
litellm==1.46.0
langgraph==0.2.35
langgraph-checkpoint-postgres
jinja2==3.1.4
pyyaml==6.0.1
httpx==0.28.1
google-cloud-aiplatform==1.67.1
psycopg[binary]==3.2.6
langfuse==2.57.0
